[Proj 1] Added the first phase
